scraper/requirements.txt
playwright==1.51.0
supabase==2.28.3
python-dotenv==1.1.0
deepl==1.21.0
openai>=1.30.0
sentry-sdk>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0